Fixed all bugs. Needs some memory clearing

+++ Sachin-BDRProblem/Basics.hpp
@@ -30,8 +30,8 @@ using namespace boost::interprocess;
 
 typedef unsigned long long  UInt64;
 typedef long long           Int64 ;
-typedef unsigned long       UInt32;
-typedef long                Int32 ;
+typedef unsigned int       	UInt32;
+typedef int                	Int32 ;
 typedef unsigned short      UInt16;
 typedef short               Int16 ;
 typedef unsigned char       UInt8;
